fix: result overview에 나오는 수치 수정
- Added a shared performance analytics utility that derives target counts, accuracy, reaction times, and data completeness directly from recorded training points.

- Updated training session completion to use the shared analytics, ensuring saved session metrics reflect the recorded tracking data.

- Reused the centralized analytics on the results page so the performance overview displays consistent, accurate metrics from the session data
